requirements.txt
streamlit
tensorflow
Pillow
pandas
numpy
opencv-python
tqdm
matplotlib
